backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
anthropic
pytesseract
pdf2image
pillow
opencv-python
prophet
pandas
scikit-learn
python-dotenv
reportlab
pydantic
numpy
langchain
langchain-anthropic
httpx
fpdf2
requests
PyMuPDF
